new update

+++ src/app/chats/chats.page.ts
@@ -11,7 +11,7 @@ import { DataService } from '../data.service';
 import { CookieService } from 'ngx-cookie-service';
 import { DOCUMENT} from '@angular/common';
 import { interval, Observable } from 'rxjs';
-import { THIS_EXPR } from '@angular/compiler/src/output/output_ast';
+import {Location} from '@angular/common'
 import { AngularFireAuth } from '@angular/fire/auth';
 
 @Component({
@@ -45,7 +45,7 @@ users$: Observable<Array< {}>>
   @ViewChild(IonContent) content: IonContent
 
 
-    constructor(private af: AngularFireAuth ,private rout: Router,private toast : ToastController,  private cookie: CookieService,  public _apiService:ApiService, public apiValue: DataService, private route: ActivatedRoute, private router: Router) {
+    constructor(private location: Location, private af: AngularFireAuth ,private rout: Router,private toast : ToastController,  private cookie: CookieService,  public _apiService:ApiService, public apiValue: DataService, private route: ActivatedRoute, private router: Router) {
     
 this.route.queryParams.subscribe(params => {
   this.values = params;
@@ -64,8 +64,14 @@ if (params && params.special){
     
   }
   onLogout(){
-    this.cookie.deleteAll();
-    this.af.signOut().then(() => this.rout.navigate(['login']));
+    if(confirm("Are you sure about Logging Out?")){
+      this.cookie.deleteAll();
+      this.af.signOut().then(() => this.rout.navigate(['login']))
+    }
+   
+    }
+    goBack(){
+      this.location.back();
     }
  /* sendMessage(){
 this.messages.push({
